gkt-water-mgmt/src/main/java/com/example/geektrust/PrintBillCommand.java
package com.example.geektrust;

import org.jetbrains.annotations.NotNull;

public class PrintBillCommand {
	final @NotNull IRateCalculator rateCalcs;

	public PrintBillCommand() {
		rateCalcs = new RateCalculators();
	}

	public Apartment parseCommand(Apartment apt) {

		Double[] total = rateCalcs.getCost(apt);

		LogUtil logger = new LogUtil();
		logger.writeFirstLine(Math.round(total[0]) + " " + Math.round(total[1]));

		// System.out.println(Math.round(total[0]) + " " + Math.round(total[1]));
		return apt;
	}
}